Extract ImageVectorPreviewPanel as reusable component
