+++ components/ir/src/main/kotlin/io/github/composegears/valkyrie/ir/Stub.kt
@@ -1,0 +1,10 @@
+package io.github.composegears.valkyrie.ir
+
+val IR_STUB = IrImageVector(
+    name = "",
+    defaultWidth = 24.0f,
+    defaultHeight = 24.0f,
+    viewportWidth = 18.0f,
+    viewportHeight = 18.0f,
+    nodes = emptyList(),
+)
