+++ idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/foundation/components/previewer/PreviewParsingError.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.editor.ui.error
+package io.github.composegears.valkyrie.ui.foundation.components.previewer
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.Image
